+++ js/utils/floatingPoint-number.js
@@ -1,0 +1,7 @@
+export const getFloatingPointNumber = (from, to, fractionDigit) => {
+  if (0 <= from && from < to && 0 <= fractionDigit) {
+    const result = Math.random() * (to - from + 1) + from;
+    return +result.toFixed(fractionDigit);
+  }
+  throw new Error('Неверное значение');
+};
